Remove specifics from shared visitor
